Refactor the name of the adapter
Before, we had the concept of an adapter. Its job was to convert data
from the API into a Ruby object. This value object represented a list of
the user's recent tracks. The object's name failed to reflect this. We
refactored the name of the adapter to better radiate its
responsibilities.